Move ResponseType to new file
